+++ Modules/SharePointDsc/Examples/Single Server/SharePoint.ps1
@@ -257,6 +257,13 @@ Configuration Example
             PsDscRunAsCredential  = $SPSetupAccount
             DependsOn             = "[SPServiceAppPool]MainServiceAppPool"
         }
+
+        SPPublishServiceApplication PublishManagedMetadataServiceApp
+        {  
+            Name           = "Managed Metadata Service Application"
+            Ensure         = "Present"
+            InstallAccount = $SetupAccount
+        }
         
         #**********************************************************
         # Local configuration manager settings